requirements.txt
pyyaml
urllib3
redis
IPy
geoip2
